--- packages/rolldown/src/rolldown.ts
@@ -1,31 +1,0 @@
-import type { InputOptions } from './options/input-options'
-import { RolldownBuild } from './rolldown-build'
-import { createWatcher } from './watch/watcher'
-import { createBundler } from './utils/create-bundler'
-import { WatchOptions } from './options/watch-options'
-import { handleOutputErrors } from './utils/transform-to-rollup-output'
-import { RolldownWatcher, WatcherEmitter } from './watch/watch-emitter'
-
-// Compat to `rollup.rollup`, it is included scan module graph and linker.
-export const rolldown = async (input: InputOptions): Promise<RolldownBuild> => {
-  return new RolldownBuild(input)
-}
-
-// Compat to `rollup.watch`
-export const watch = (input: WatchOptions): RolldownWatcher => {
-  const emitter = new WatcherEmitter()
-  createWatcher(emitter, input)
-  return emitter
-}
-
-/**
- * @description
- * This is an experimental API. It's behavior may change in the future.
- * Calling this API will only execute the scan stage of rolldown.
- */
-export const experimental_scan = async (input: InputOptions): Promise<void> => {
-  const { bundler, stopWorkers } = await createBundler(input, {})
-  const output = await bundler.scan()
-  handleOutputErrors(output)
-  await stopWorkers?.()
-}
